// projects/services/src/lib/models/enums/PageInkPosition.ts
export enum PageInkPosition {
  Home = 15,
  About = 112,
  Search = 213,
  Faq = 380,
  PlasmaDevice = 502,
  TrainingPrograms = 285,
  Contact = -150,
  GetStarted = -150
}
export enum PageInkWidth {
  Home = 60,
  About = 131,
  Search = 136,
  Faq = 146,
  PlasmaDevice = 135,
  TrainingPrograms = 175,
  Contact = 1,
  GetStarted = 1
}